Add Alias For Employee & Payroll For Their IDs
Add an alias for Employee & Payroll for unique ids fields outside the
standard table ids because it reads better in console and code with no
duplication of the model name. So `employee.employee_id` is now
`employee.identifier` as it flows better, making it easier to
understand that this is the id to use instead of the one in the table.