Update Payrolls CTRL to Prevert Use Non-CSV Files
Update the Payrolls controller to prevent non-CSV files since only CSVs
are allowed to be used to add payrolls would enforce this rule on any
platform, not just the website. Thus, each application is not required
to enforce this rule locally. However, for speed of getting back to the
user as quickly on which types are allowed. It is encouraged this is to
implement locally.