Add `successful_payrolls` to Track Good Payrolls
Add `successful_payrolls` to track the payrolls that can add. However,
there is the ease of use and the user not to rename the file after
fixing whichever row has broken. Instead, all of the rows must be good
to be all under the same Payroll Id.

Thus, the user can track which Payrolls have complete and which ones
have not without having to remember that some of the Payrolls had errors
in them and are actually in another Payroll Id.